+++ README.md
@@ -1,2 +1,11 @@
 # YP_diplom.github.io
+
 Дипломный проект курса фронтенд-разработчика от яндекс практикума.
+
+Для запуска проекта клонируем его командой: git clone https://github.com/Artem74Arp/YP_diplom.github.io.git
+
+Для установки всех используемых в проекте npm пакетов используем команду: npm i
+
+Для сборки проекта используем команды: npm run build и npm run dev, вуаля вы прекрасны:)
+
+[Ссылка на проект](https://artem74arp.github.io/YP_diplom.github.io/)
