--- src/about/index.js
@@ -1,4 +1,0 @@
-import "./about.css";
-
-import sliderSettings from "../blocks/commits/__slider/slidersettings"
-sliderSettings();
