+++ src/analytics/analytics.html
@@ -7,18 +7,16 @@
   <meta name="author" content="Артем Петров"/>
   <meta name="description" content="NewsAnalyzer - это отличный сервис для того, чтобы быть в курсе событий и получать аналитику новостей."/>
   <title>Analytics</title>
-  <link rel="stylesheet" href="<%= htmlWebpackPlugin.files.chunks.main.css %>">
-  <link href="<%=htmlWebpackPlugin.files.favicon%>" type="image/png" rel="shortcut icon">
 </head>
 <body>
   <div class="root">
     
     <header class="header">
-      <div class="header__container">
+      <nav class="header__nav">
         <a href="./index.html" class="link header__logo header__logo_theme_dark">NewsAnalyzer</a>
         <a href="./index.html" class="link header__link-main header__link-main_theme_dark">Главная</a>
         <a href="./about.html" class="link header__link-about header__link-about_theme_analytics">О проекте</a>
-      </div>
+      </nav>
     </header>
 
     <main class="content">
@@ -116,7 +114,6 @@
   
   </div>
 
-  <script src="<%= htmlWebpackPlugin.files.chunks.analytics.entry %>"></script>
 
 </body>
 </html>
