+++ src/index.css
@@ -9,5 +9,3 @@
 @import url("./blocks/about/about.css");
 @import url("./blocks/footer/footer.css");
 @import url("./blocks/hidden/hidden.css");
-
-
