+++ src/index.html
@@ -7,8 +7,6 @@
   <meta name="author" content="Артем Петров"/>
   <meta name="description" content="NewsAnalyzer - это отличный сервис для того, чтобы быть в курсе событий и получать аналитику новостей."/>
   <title>NewsAnalyzer</title>
-  <link rel="stylesheet" href="<%= htmlWebpackPlugin.files.chunks.main.css %>">
-  <link href="<%=htmlWebpackPlugin.files.favicon%>" type="image/png" rel="shortcut icon">
 </head>
 <body>
   <div class="root">
@@ -121,8 +119,5 @@
   
   </div>
 
-  <script src="<%= htmlWebpackPlugin.files.chunks.main.entry %>"></script>
-
 </body>
 </html>
-
